requirements-production.txt
# Production Requirements for DidactAI
# Core Django Dependencies
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0

# Database - PostgreSQL for production
psycopg2-binary==2.9.9

# Authentication & Security
django-allauth==0.57.2
django-crispy-forms==2.1
crispy-tailwind==0.5.0

# File Handling & Storage
Pillow==10.3.0
django-storages==1.14.2
boto3==1.34.51

# Document Processing - Production optimized versions
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
pytesseract==0.3.10

# AI & NLP Libraries - Stable versions
langchain==0.1.13
langchain-google-genai==1.0.3
google-generativeai==0.4.1
transformers==4.39.3
torch==2.2.2
sentence-transformers==2.6.1
langdetect==1.0.9

# Export & PDF Generation
reportlab==4.1.0
weasyprint==61.2

# Production Server & Caching
gunicorn==21.2.0
whitenoise==6.6.0
redis==5.0.3
celery==5.3.6

# Utilities
requests==2.31.0
python-magic-bin==0.4.14  # Windows compatible
zipfile36==0.1.3

# Monitoring & Error Tracking
sentry-sdk==1.45.0

# Internationalization
django-rosetta==0.9.9

# Security & Performance
django-ratelimit==4.1.0
django-csp==3.7